Fix interfaces deserialization on .NET Core
