+++ src/ServiceStack.Text/ServiceStack.Text.Android.csproj
@@ -264,6 +264,8 @@
     <Compile Include="XmlSerializer.cs">
       <SubType>Code</SubType>
     </Compile>
+    <Compile Include="DynamicProxy.cs" />
+    <Compile Include="PclExport.NetStandard.cs" />
   </ItemGroup>
   <!--ItemGroup,Content-->
   <!--ItemGroup,None-->
