+++ src/ServiceStack.Text/ServiceStack.Text.SL5.csproj
@@ -292,6 +292,8 @@
       <SubType>Code</SubType>
     </Compile>
     <Compile Include="ParseAsType.cs" />
+    <Compile Include="DynamicProxy.cs" />
+    <Compile Include="PclExport.NetStandard.cs" />
   </ItemGroup>
   <!--ItemGroup,Content-->
   <!--ItemGroup,None-->
