+++ src/ServiceStack.Text/ServiceStack.Text.Signed.csproj
@@ -324,6 +324,8 @@
       <SubType>Code</SubType>
     </Compile>
     <Compile Include="ParseAsType.cs" />
+    <Compile Include="DynamicProxy.cs" />
+    <Compile Include="PclExport.NetStandard.cs" />
   </ItemGroup>
   <ItemGroup>
     <BootstrapperPackage Include="Microsoft.Net.Client.3.5">
