+++ src/ServiceStack.Text/ServiceStack.Text.csproj
@@ -318,6 +318,8 @@
       <SubType>Code</SubType>
     </Compile>
     <Compile Include="ParseAsType.cs" />
+    <Compile Include="DynamicProxy.cs" />
+    <Compile Include="PclExport.NetStandard.cs" />
   </ItemGroup>
   <ItemGroup>
     <BootstrapperPackage Include="Microsoft.Net.Client.3.5">
