+++ tests/ServiceStack.Text.Tests/JsonTests/ContractByInterfaceTests.cs
@@ -48,9 +48,6 @@ namespace ServiceStack.Text.Tests.JsonTests
         }
 
         [Test]
-#if NETCORE
-	[Ignore("Throws StackOverflowException")]
-#endif
         public void Should_be_able_to_deserialise_based_on_an_interface_with_no_concrete()
         {
             using (JsConfig.With(preferInterfaces: true))
